Nathan

+++ app/src/main/java/com/operatorsapp/adapters/ShiftLogSqlAdapter.java
@@ -20,6 +20,7 @@ import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import com.example.common.Event;
 import com.operatorsapp.BuildConfig;
 import com.operatorsapp.R;
 import com.operatorsapp.application.OperatorApplication;
@@ -28,7 +29,6 @@ import com.operatorsapp.utils.ReasonImage;
 import com.operatorsapp.utils.ReasonImageLenox;
 import com.operatorsapp.utils.TimeUtils;
 import com.ravtech.david.sqlcore.DatabaseHelper;
-import com.ravtech.david.sqlcore.Event;
 
 import java.util.ArrayList;
 
@@ -47,13 +47,13 @@ public class ShiftLogSqlAdapter extends CursorRecyclerViewAdapter<RecyclerView.V
     private OnStopClickListener mOnStopClickListener;
     public LayoutInflater inflater;
     private boolean mIsSelectionMode;
-    private ArrayList<Integer> mSelectedEvents;
+    private ArrayList<Float> mSelectedEvents;
     private ArrayList<Event> mUpdatedAlarms;
     private int mFirstStopEventPosition;
 
     public ShiftLogSqlAdapter(Context context, Cursor cursor, boolean closedState, int closeWidth,
                               OnStopClickListener onStopClickListener, int openWidth, int height,
-                              boolean selectMode, ArrayList<Integer> selectedEvents) {
+                              boolean selectMode, ArrayList<Float> selectedEvents) {
         super(cursor);
         mContext = context;
         mClosedState = closedState;
@@ -67,7 +67,7 @@ public class ShiftLogSqlAdapter extends CursorRecyclerViewAdapter<RecyclerView.V
         mFirstStopEventPosition = getItemCount();
     }
 
-    public void setSelectedEvents(ArrayList<Integer> selectedEvents) {
+    public void setSelectedEvents(ArrayList<Float> selectedEvents) {
         mSelectedEvents = selectedEvents;
     }
 
@@ -172,7 +172,7 @@ public class ShiftLogSqlAdapter extends CursorRecyclerViewAdapter<RecyclerView.V
                 cursor.getString(cursor.getColumnIndex(DatabaseHelper.KEY_TIME)),
                 cursor.getString(cursor.getColumnIndex(DatabaseHelper.KEY_END_TIME)),
                 cursor.getInt(cursor.getColumnIndex(DatabaseHelper.KEY_GROUP_ID)),
-                cursor.getInt(cursor.getColumnIndex(DatabaseHelper.KEY_EVENT_ID)),
+                cursor.getFloat(cursor.getColumnIndex(DatabaseHelper.KEY_EVENT_ID)),
                 cursor.getInt(cursor.getColumnIndex(DatabaseHelper.KEY_PRIORITY)),
                 cursor.getString(cursor.getColumnIndex(DatabaseHelper.KEY_GROUP_E_NAME)),
                 cursor.getString(cursor.getColumnIndex(DatabaseHelper.KEY_GROUP_L_NAME)),
@@ -236,7 +236,7 @@ public class ShiftLogSqlAdapter extends CursorRecyclerViewAdapter<RecyclerView.V
             }
 
             if (mSelectedEvents != null) {
-                for (Integer event1 : mSelectedEvents) {
+                for (Float event1 : mSelectedEvents) {
                     if (event.getEventID() == event1) {
                         event.setChecked(true);
                     }
@@ -485,7 +485,7 @@ public class ShiftLogSqlAdapter extends CursorRecyclerViewAdapter<RecyclerView.V
         }
     }
 
-    private void validateDialog(final int eventID) {
+    private void validateDialog(final float eventID) {
 
         AlertDialog.Builder builder = new AlertDialog.Builder(mContext);
         LayoutInflater inflater = (LayoutInflater) mContext.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
